feat(db): split misc.ts final batch — packetLog + keyRepair + teardown (Task 3.1 PR3)
Move the remaining two domains out of MiscRepository and delete misc.ts:

- packetLog.ts: PacketLogRepository (19 public methods + 3 private helpers +
  PacketLogFilterOptions interface); includes analytics, sync twins, raw-SQL
  joins for node-name resolution
- keyRepair.ts: KeyRepairRepository (7 SQLite-only sync methods, verbatim
  move; async parity deferred to Task 3.2)

Teardown:
- Delete misc.ts, misc.packetlog.test.ts (replaced by packetLog.test.ts),
  misc.mapprefs.test.ts (duplicate of mapPreferences.test.ts from PR2)
- Remove MiscRepository import, miscRepo field, and get misc() getter from
  DatabaseService; add packetLogRepo/keyRepairRepo fields + getters
- Repoint all this.misc.X → this.packetLog.X and this.miscRepo!.X → per-domain
  repo calls throughout database.ts
- Update database.service.test.ts and database.purgeAllNodes.test.ts mocks
- Update index.ts barrel: export PacketLogRepository/PacketLogFilterOptions/
  KeyRepairRepository; remove MiscRepository
- misc.test.ts → end-state assertion (MiscRepository unexported, misc.ts gone,
  new repos exported)
- Lint baseline updated net-zero: misc.ts -44 any violations,
  packetLog.ts +33 + keyRepair.ts +11 = 0 net change
- Update REMEDIATION_EPIC.md: check off 3.1, add phase log entry

Verification: typecheck clean; typecheck:tests 283 (baseline); lint:ci exit 0;
build succeeded; suite success=true, 8472 passed, 0 failed, 0 suitesFailed.
